transcode: improve low-bitrate profiles bitrate slightly for better playback (#216)
* transcode: improve low-bitrate profiles bitrate slightly for better playback experience

When the input video's height or bitrate combo is too low to meet
the ABR ladder, the low-bitrate profile is used in addition to the
input video's profile to have some ABR during playback. The low-bitrate
profile 1/3's the input video profile resulting in pretty low quality
video. Increase this to be 75% of the input video's resolution so that
playback still feels good without experiencing artifacts.

* Switch to 50% quality

Co-authored-by: Thom Shutt <thomshutt@users.noreply.github.com>