handlers,pipeline: Move segmenting target URL logic to mist pipeline
Found a bug on the background mist job: Since the coordinator only
changes the TargetURL when the job is in background (line 206),
the mist job will actually not pick up that change since it uses
the segmenting URL instead.

Fix this the right way, by moving the creation of the segmeting URL
to the mist pipeline. It is only used there.

The only small issue is that it is still a "global" field in the job
info, so I added a comment that it's only used in mist for now. We
can do something more sophisticated is this grows (like a "handler data"
inside the job info struct that can be of any type)