clients/mediaconvert: transcode with source url if download and upload to S3 fails (#327)
Certain file servers may close the connection when catalyst-api is in
the process of downloading the source file - an example is IPFS based
links where the gateway and pinning servers may be different and lead to
erratic connection issues. Instead of giving up, catalyst-api will now
attempt to transcode using the source URL directly since the AWS/MediaConvert
API has built-in support for retries. This will ensure more successful
transcode sessions albeit missing a copy of the source input file which
is only required for debugging.

Note: a follow-up PR will be used to implement download-manager style
retries in catalyst-api.