probe: refactor ffprobe logic and populate metadata
* The ffprobe logic is being refactored to abstract away all ffprobe
specific bits. The video.Probe() function will return an InputVideo{}
that is already formatted and ready to use. This will make it easier to
integrate probing logic at input or output sides much more easily
throughout the code base.

* The probed output MP4 files is used to return metadata on these MP4
files to Studio which will be used by the SDK to optimize playback for
short-form video content.