Implement unified input copying and probing (#463)
This covers moving the input file copying and probing from the mediaconvert pipeline to the Coordinator so that we have access to the probe data ahead of any of the pipeline specific code. The object store copy of the input file can then be used as the input for mist as well as mediaconvert.

The bulk of the change lives in pipeline/coordinator.go, along with some refactoring in clients/input_copy.go. There are then a fair amount of changes to move the copy & probe related tests from clients/mediaconvert_test.go to pipeline/coordinator_test.go

* implement unified input copying and probing

* move copy and probe to before pipeline start

* fix unit tests

* separate out transfer directory from mist source

* address review comments

- add some unit tests to cover more probing failures
- reduced the presigning validity time - accidentally left this much
higher
- removed the presigned url from the log lines to avoid leaking it