transcode: skip mp4 output for low-bitrate profiles
When the input video's res/bitrate combo is too low for the default ABR
ladder, an additional low-bitrate profile with 50% of input bitrate is
generated for ABR HLS based playback. For MP4s, only a single MP4
matching the default ABR is required so skip MP4 output for low bitrate
profiles.