chore: fix linting rules
Some of these are real fixes.

For mapic, which is deprecated, doing things like rewriting `ioutil` usage
would introduce more risk than it would fix. So we just disable linting there.
Things can be incrementally improved as we refactor it out into other projects.

The last category is writing to HTTP responses, which is a thing that the Go
standard library does all the time.
Context: https://discord.com/channels/423160867534929930/1072280408777039903/1090740805565694042