mapic: Fix parsing of human_issues field
It is undocumented and I assumed it was a string just like
the issues. Turns out it is a string array.

If I just changed the types it would make the parsing logic
much more convoluted, as the array is actually parsed as a
[]interface{} and I'd need to check each field. So I changed
it once again, making yet another parsing of the input object.