mist-cleanup: fix bash globbing
With the double-quotes, bash did not split the variable on special
characters and used it as a whole string. Without the double-quotes,
bash splits the variable on special characters. Mist's /dev/shm seem to
have some special characters that was causing failures by inadvertently
deleting files that were not supposed to be deleted.