Fix recordings
We want to pass the plain HTTP input URL through for recordings because this is the simplest way to allow mediaconvert to read the input (in case of livepeer pipeline failures). Some parts of the code were expecting an OS URL though so this is what is fixed in this PR, to allow plain HTTP URLs to work too.