transcode: generate mp4s for only the highest bitrate rendition for HLS (#769)
input

It was discovered that pods were hitting OOM errors when long running
livestreams and their corresponding recordings were being processed. In
the recording generation stage, the ts segments are buffered in memory
and then written to disk to generate the mp4. This poses an issue for
livestreams that are several hours in duration. Additionally, it doesn't
make sense to generate mp4s for all renditions for HLS inputs.

This commit updates the logic such that an mp4 will be generated for
only the highest bitrate profile.