input_copy: "manually" parse duration of HLS streams when ffprobe fails
For some HLS streams, the duration cannot be detected correctly and
ffprobe reports "N/A" as duration. This was evident in long running
livestreams that were 100s of hours in length. In such cases, we parse
the manifest file using the m3u library to manually determine the
duration by adding up segment lengths as reported by the EXTINF tags.