clip: add scaffolding to accept clipping requests
This adds additional parameters to the vod endpoint to accept clipping
start/end times. If the fields are set with valid values, the input
manifest will be clipped and then a new clipped manifest will be used in
the VOD pipeline to generate assets as usual. The logic for clipping the
actual manifest will be done as a follow-on PR.