clip: add parameter to set clip output dir in upload request
This output dir will typically be the same bucket and root folder where
the recording manifest resides. It will contain the clipped segments
(start + end) and the clip manifest. The transcoded VOD assets of the
clips will reside in the usual location set by the hls target output
field. Currently both clips and transcoded assets are output to the same
directory which is incorrect since we want them in different buckets.