transcode: add logic to force a session re-init
When a sequence of segments has a discontinuity (e.g. encoded
differently), the transcoder at T must be re-initialized so that
transcode operations can be completed correctly. The T takes some
liberties and assumes a sequence of segments are encoded the same way
and thus, ends up not re-initializing certain parts of the pipeline
(e.g. demuxer, etc) to save time and improve efficiency. This can result
in corrupted segments with two video tracks instead of one audio and one
video track.

For clipping, the first segment (index=0) is encoded differently from
the rest of the file. As such, a session reinit must be forced at the T
using a header set in the request to B.