capacity: use separate limits for regular vod vs clipping vod jobs
Clipping requests may arrive in a spiky fashion and short clips (under 2 mins)
are processed typically within ~60s (measured from when a clipping
request arrives at catalyst to when the final mp4s are returned).

Separate limits for clip-vod and regular-vod will be used to allow
clipping to scale when there's increased usage.

The limits here in this change are still somewhat arbitrary and will
need to be tuned as we load test.