transcode: memory usage optimization
Transcoded segments are held in memory until the MP4 generation stage
(after transcode stage is completed). This results in large usage in RAM
when processing very long recordings that are several hours in length.
This can occasionally result in kube pods hitting OOM.

To optimize this behavior, this PR makes the following changes:

* Create a buffered channel that holds transcoded segments
* Launch a disk-writing goroutine that consumes the buffered channel and
  writes segments to disk.
* Delete the byte array of the segment's data in hopes the garbage
  collector releases it from memory.
* Switch all MP4 generation to use stream-based concatenation instead of
  file-based concatenation.