transmux: use file-based concatenation for non-clips
Certain input files have a/v sync issues in the output mp4s when they
are played in the browser -- adjusting the playhead fixes the sync
issues. This seems to only occur on browsers and not standalone players
like vlc and qt. It was found that using file based concatenation fixes
this issue. The root cause is under investigation but likely caused by
incorrect timestamps between audio and video tracks. This commit updates
the mp4 generation logic to use stream-based concat for only clips and
use file-based concat for non-clips.