transcode: drop first segment if a/v tracks have a large offset
For HLS input (recordings), sometimes the first segment can experience a
large delta between the audio and video track's start time (in seconds).
This is mainly because when a streamer begins the stream, the user's
hardware config and OBS settings might cause the video track to start
later than the audio (or vice versa). These segments can be transcoded
but muxing them into an MP4 will fail. For these cases, the first segment
should be dropped from the transcode payload.