transcode: handle a/v offset case for short recordings
If the first segment is dropped due to a large offset between the a/v
track and it happens to be the only segment in the stream (e.g. very
short recording), then flag it as an error that will be excused instead
of resulting in a panic when attempting to access an empty slice.