manifest: update manifest re-ordering to use 960p as threshold
To optimize playback and TTFF, the output transcode ladder is being
updated to use 960p as the threshold i.e. any width/height above 960p
will be re-ordered such that it appears second in the list. This is
optimal for typical resolutions. For example, if the highest quality is
1920x1080, then it will likely get swapped with 1280x720 in the transcode ladder
ordering. Similarly, if the ladder has a highest quality of 1280x960, it
will get swapped with the next likely occuring resolution like 1280x720.