mist-balancer/geolocation: use tag_adjust weight of 0 for pull ingest requests
When Studio calls catalyst-api to begin a pull ingest, the lat/lon is
provided as a query param. When this query param is detected and valid,
the tag_adjust setting used to prefer the current node that was picked
using DNS rules *should not* apply. This is because the Studio request
can come from any region and the DNS based routing might cause the
request to be processed in an incorrect far away region. In this case,
we only care about the lat/lon passed in as a query param which will get
passed to MistUtilLoad and get geolocated with a heavier global 'geo'
weight setting (set in livepeer-infra).