clip: add more logging to debug failed clips
Clips are failing reporting 'start time specified exceeds duration of
manifest'. This may happen for webrtc streams where the clip request
comes in but the segments and manifest hasn't been written yet to S3. I
n this case, the start time would fall past the latest segment and this
error case would be hit -- however, we have several retries for this
exact scenario.

Based on logs, there are 10 retries every 5 seconds in
addition to 3 studio retries - this should be sufficient for the latest
segments to be written to S3 and the clip request would eventually
succeed. This is not happening so we're adding the additional logging to
ensure latest manifests are being downloaded on retry attempts.